Dockerfiles for Ch 04
Since Mac (and to a lesser extent Windows, though it has WSL) cannot work with epoll directly, Dockerfiles,
Docker compose yml, and simple run scripts have been created so that users who have docker can run and test the ch04 examples on their
system.

A docker file has been created for delayserver as well and docker
compose is used to run an epoll container and a delay server
container and coordinate the network between them.

Some slight modification has been made to the epoll code to allow
for an optional argument to be passed in which is the name of the
host. When running the examples without docker both are running on
"localhost". However when running docker containers epoll and
delayserver are in separate containers, so epoll needs to transmit
to the delayserver container (conveniently named "delayserver") instead of localhost.

Debian slim rust containers have been used as they are a convenient
and fairly small container for these kinds of rust projects.

README.md's have been updated to note that it is now possible to pass an optional argument for the hostname to connect to, which will default to "localhost" if no argument is passed.